+++ server/utils/map_creator/map.png
