+++ server/utils/pathfinding/astar.py
@@ -18,7 +18,7 @@ class PriorityQueue:
 
 def heuristic(node, goal):
     # Other types of heuristics can be possibly be used
-    value = vector.Vector(node.position, goal.position).magnitude
+    value = vector.Vector(node, goal).magnitude
     return value
